Workaround for rpath issues
Two things going on that are a bit odd/suspicious:
1) By default, qmake adds a -rpath with the full sysroot path and Qt installation
prefix. However, this doesn't actually seem to work and we need to add one with
only the Qt prefix (despite the --sysroot option)
2) When -dont-process is passed to configure, for some reason the --sysroot option
is no longer passed to the compiler.

Need to investigate both of these to see if they're bugs or features