Reworked the scripts to accept any qtbase
Rather than require that the script use the qtbase from a submodule the
configure-qt.sh script accepts a path to qtbase so that developers can
use whatever copy they'd like.

And as the name implies configure-qt.sh (as opposed to the old build-qt.sh)
only configures qtbase. It is up to the developer to build qtbase as well
as any other Qt modules.